Move BasketProvider and GridContainer
Basket functions now lie in BasketProvider
GridContainer is layout
