+++ frontend/src/App.tsx
@@ -1,13 +1,16 @@
 import React from 'react';
 import './App.css';
 import './index.css';
-import ShoppingHandler from './ShoppingHandler';
+import {BasketProvider} from './BasketProvider';
+import {GridContainer} from './GridContainer';
 
 function App() {
 	return (
 		<div>
 			<h2>SHOPshop</h2>
-			<ShoppingHandler/>
+			<BasketProvider>
+				<GridContainer/>
+			</BasketProvider>
 			<h4>SHOPshop</h4>
 		</div>
 	);
